bullet3: use double precision
Summary:
Double precision Bullet3 is now a hard requirement for OpenMW 0.47.0.
Note: while OpenMW is hardcoded to use specific precision (float in pre-0.47 releases and double in 0.47 and future releases), the build system of Godot adapts to precision of the provided build, so it should rebuild normally.

Test Plan:
1. Built and tested `openmw` 0.47.0 against it.
2. Rebuilt `godot`.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Tags: #requires_rebuilds

Differential Revision: https://dev.getsol.us/D12044